+++ lib/PhpParser/Node/Scalar/LNumber.php
@@ -62,6 +62,11 @@ class LNumber extends Scalar
             throw new Error('Invalid numeric literal', $attributes);
         }
 
+        // Strip optional explicit octal prefix.
+        if ('o' === $str[1] || 'O' === $str[1]) {
+            $str = substr($str, 2);
+        }
+
         // use intval instead of octdec to get proper cutting behavior with malformed numbers
         $attributes['kind'] = LNumber::KIND_OCT;
         return new LNumber(intval($str, 8), $attributes);